// src/app/types/tickets/Ticket.ts
export interface Ticket {
    Name: string;
    Price: number;
    Mapping: string;
    SeatTaken: string;
    LoyaltyDiscount: number;
    HideEpay: string;
    Tax1: number;
    Tax2: number;
    Tax3: number;
    Tax4: number;
    MealPrice: number;
    OtsgPrice: number; // other surcharges
    ProdG08: string;
    Quantity: number;    
  }